chore: relocate import-manifest kind to its provider and defer unused ImportManifest accessor
