+++ cli/internal/project/importmanifest/provider_test.go
@@ -19,7 +19,7 @@ func manifestSpec(workspaceID string, ids ...specs.ImportIds) *specs.Spec {
 		})
 	}
 	return &specs.Spec{
-		Kind:    specs.KindImportManifest,
+		Kind:    KindImportManifest,
 		Version: specs.SpecVersionV1,
 		Spec: map[string]any{
 			"workspaces": []any{
@@ -36,34 +36,11 @@ func TestProvider_TypeSurfaces(t *testing.T) {
 	t.Parallel()
 
 	p := New()
-	assert.Equal(t, []string{specs.KindImportManifest}, p.SupportedKinds())
+	assert.Equal(t, []string{KindImportManifest}, p.SupportedKinds())
 	assert.Nil(t, p.SupportedTypes())
 	assert.Equal(t, []rules.MatchPattern{
-		rules.MatchKindVersion(specs.KindImportManifest, specs.SpecVersionV1),
+		rules.MatchKindVersion(KindImportManifest, specs.SpecVersionV1),
 	}, p.SupportedMatchPatterns())
-}
-
-func TestProvider_LoadSpec_Appends(t *testing.T) {
-	t.Parallel()
-
-	p := New()
-	require.NoError(t, p.LoadSpec("a.yaml", manifestSpec("ws-1",
-		specs.ImportIds{URN: "source:src-1", RemoteID: "remote-1"})))
-	require.NoError(t, p.LoadSpec("b.yaml", manifestSpec("ws-2",
-		specs.ImportIds{URN: "destination:dst-1", RemoteID: "remote-2"})))
-
-	assert.Equal(t, &specs.WorkspacesImportMetadata{
-		Workspaces: []specs.WorkspaceImportMetadata{
-			{
-				WorkspaceID: "ws-1",
-				Resources:   []specs.ImportIds{{URN: "source:src-1", RemoteID: "remote-1"}},
-			},
-			{
-				WorkspaceID: "ws-2",
-				Resources:   []specs.ImportIds{{URN: "destination:dst-1", RemoteID: "remote-2"}},
-			},
-		},
-	}, p.ImportManifest())
 }
 
 func TestProvider_LoadSpec_RejectsMalformed(t *testing.T) {
@@ -75,28 +52,6 @@ func TestProvider_LoadSpec_RejectsMalformed(t *testing.T) {
 	})
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "parsing import-manifest bad.yaml")
-}
-
-func TestProvider_ImportManifest_NilWhenEmpty(t *testing.T) {
-	t.Parallel()
-
-	assert.Nil(t, New().ImportManifest())
-}
-
-func TestProvider_ImportManifest_ReturnsCopy(t *testing.T) {
-	t.Parallel()
-
-	p := New()
-	require.NoError(t, p.LoadSpec("a.yaml", manifestSpec("ws-1",
-		specs.ImportIds{URN: "source:src-1", RemoteID: "remote-1"})))
-
-	// Mutating the returned slice must not corrupt provider state.
-	got := p.ImportManifest()
-	got.Workspaces = append(got.Workspaces, specs.WorkspaceImportMetadata{WorkspaceID: "ws-injected"})
-
-	again := p.ImportManifest()
-	assert.Len(t, again.Workspaces, 1)
-	assert.Equal(t, "ws-1", again.Workspaces[0].WorkspaceID)
 }
 
 func TestProvider_LoadLegacySpec_Unsupported(t *testing.T) {
